requirements.txt
streamlit
pandas
networkx
matplotlib
torch
torchvision
pillow